Move channel_options under grpc_internals
